requirements.txt
Flask==1.0.2
Flask-SQLAlchemy-2.3.2
Flask-Session-0.3.1
